+++ airbyte-integrations/connectors/source-qualaroo/source_qualaroo/spec.json
@@ -4,7 +4,11 @@
     "$schema": "http://json-schema.org/draft-07/schema#",
     "title": "Qualaroo Spec",
     "type": "object",
-    "required": ["token", "key", "start_date"],
+    "required": [
+      "token",
+      "key",
+      "start_date"
+    ],
     "additionalProperties": true,
     "properties": {
       "token": {
@@ -24,13 +28,15 @@
         "title": "Start Date",
         "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}.[0-9]{3}Z$",
         "description": "UTC date and time in the format 2017-01-25T00:00:00Z. Any data before this date will not be replicated.",
-        "examples": ["2021-03-01T00:00:00.000Z"]
+        "examples": [
+          "2021-03-01T00:00:00.000Z"
+        ]
       },
       "survey_ids": {
         "type": "array",
         "items": {
           "type": "string",
-          "pattern": "^[0-9a-fA-F]{24}$"
+          "pattern": "^[0-9]{1,8}$"
         },
         "title": "Qualaroo survey IDs",
         "description": "IDs of the surveys from which you'd like to replicate data. If left empty, data from all surveys to which you have access will be replicated."
@@ -42,7 +48,14 @@
     "oauth2Specification": {
       "rootObject": [],
       "oauthFlowInitParameters": [],
-      "oauthFlowOutputParameters": [["token"], ["key"]]
+      "oauthFlowOutputParameters": [
+        [
+          "token"
+        ],
+        [
+          "key"
+        ]
+      ]
     }
   }
 }
